Rename map theme provider to user preferences
- Move map theme state into `user_preferences_provider`
- Update imports across app and tests
