+++ lib/widgets/text_editing_shortcut_scope.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:icarus/const/shortcut_info.dart';
-import 'package:icarus/providers/map_theme_provider.dart';
+import 'package:icarus/providers/user_preferences_provider.dart';
 
 class TextEditingShortcutScope extends ConsumerWidget {
   const TextEditingShortcutScope({
